+++ snuba/datasets/storages/transactions.py
@@ -83,7 +83,7 @@ columns = ColumnSet(
         (
             "spans",
             Nested(
-                [("op", String()), ("group", UInt(64)), ("exclusive_time", Float(64))]
+                [("op", String()), ("group", UInt(64)), ("exclusive_time", Float(32))]
             ),
         ),
         ("partition", UInt(16)),
